feat(lang): null-safety migrations; tests passing

+++ example/supporting_files/parameters/power_of_two.parameter.dart
@@ -3,8 +3,12 @@ import 'dart:math';
 
 class PowerOfTwoParameter extends CustomParameter<int> {
   PowerOfTwoParameter()
-      : super('POW', RegExp(r'([0-9]+\^[0-9]+)', caseSensitive: true), (input) {
-          final parts = input.split('^');
-          return pow(int.parse(parts[0]), int.parse(parts[1]));
-        });
+      : super(
+          'POW',
+          RegExp(r'([0-9]+\^[0-9]+)', caseSensitive: true),
+          (input) {
+            final parts = input.split('^');
+            return pow(int.parse(parts[0]), int.parse(parts[1])) as int;
+          },
+        );
 }
